requirements.txt
streamlit>=1.30.0
easyocr>=1.7.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
rapidfuzz>=3.0.0
numpy>=1.24.0